Created NoTasksView, Added check and uncheck function for ItemsModel, updated ListViewModel and ItemsModel

+++ DailyDo/Views/ListView.swift
@@ -14,18 +14,27 @@ struct ListView: View {
     
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
+        if listViewModel.items.count == 0 {
+            NoItemsView()
+        } else {
+            List {
+                ForEach(listViewModel.items) { item in
+                    ListRowView(item: item)
+                        .onTapGesture {
+                            withAnimation(.linear) {
+                                listViewModel.updateItem(item: item)
+                            }
+                        }
+                }
+                .onDelete(perform: listViewModel.deleteItem)
+                .onMove(perform: listViewModel.moveItem)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+            .listStyle(PlainListStyle())
+            .navigationTitle("DailyDo App 📝")
+            .navigationBarItems(
+                leading: EditButton(),
+                trailing: NavigationLink("Add", destination: AddView()))
         }
-        .listStyle(PlainListStyle())
-        .navigationTitle("DailyDo App 📝")
-        .navigationBarItems(
-            leading: EditButton(),
-            trailing: NavigationLink("Add", destination: AddView()))
     }
 }
